--- doctorSide/Location sharing/README.md
@@ -1,1 +1,0 @@
-# https://gamma.app/doc/Location-Sharing-Page-b23ec9wyzolz0u4
