+++ doctorSide/docPage.css
@@ -1,0 +1,34 @@
+:root{
+  --bg-clr : #212032;
+}
+*{
+  box-sizing: border-box;
+  margin: 0px;
+  padding: 0px;
+}
+body {
+  /* background-image: url(https://i1.wp.com/dl.img-news.com/dl/img/s0/dl/2020/09/1471701-559173709.jpeg?fit=%2C&ssl=1); */
+  background-size: cover;
+  background-color: var(--bg-clr);
+  color: #fff;
+}
+form {
+  background-color: var(--bg-clr);
+  color: aliceblue;
+}
+#error {
+  background-color: red;
+  color: white;
+  font-size: 21px;
+  padding-left: 5px;
+}
+.btn{
+  margin-top: 2rem;
+  width: 15rem;
+}
+.container{
+  color: white;
+}
+footer{
+  height: 5rem;
+}
