+++ doctorSide/docPage.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/css/bootstrap.min.css"
+        integrity="sha384-zCbKRCUGaJDkqS1kPbPd7TveP5iyJE0EjAuZQTgFLD2ylzuqKfdKlfG/eSrtxUkn" crossorigin="anonymous">
+    <title>Doc page</title>
+    <script src="https://kit.fontawesome.com/8beace882b.js" crossorigin="anonymous"></script>
+    <link rel="stylesheet" href="docPage.css">
+
+
+    <script src="https://code.jquery.com/jquery-3.6.0.min.js"
+        integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin="anonymous"></script>
+    <!-- <link rel="stylesheet" href="./map.css"> -->
+
+    <script src="../Firebase/config.js" defer type="module"></script>
+</head>
+
+<body>
+
+    <h1 class="text-center text-uppercase bg-danger text-light rounded px-5">Doc page</h1>
+    <div class="container description-survey">
+        <p class="lead">Hi <span id="userName">Name</span><br> To assure you the best and accurate First Aid , Do fill
+            out the form correctly
+            <hr class="border border-dark">
+        </p>
+
+    </div>
+
+    <div class="row d-flex">
+        <div class="col-md-8 col-12">asdfasdfsdf</div>
+        <div class="col-md-4 col-12">
+            <button class="btn btn-primary">Call patient</button><br>
+            <button class="btn btn-primary">Get Patient info</button>
+        </div>
+
+    </div>
+
+
+
+
+
+
+    </div>
+    <footer></footer>
+    <script src="https://cdn.jsdelivr.net/npm/jquery@3.5.1/dist/jquery.slim.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js"></script>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/js/bootstrap.min.js"></script>
+    <script src="docPage.js"></script>
+    <!-- <script src="map.js"></script> -->
+</body>
+
+</html>
